Move `CThunkBase.h` from `platform/source/` to `platform/internal/`
