+++ platform/CThunk.h
@@ -39,7 +39,7 @@
 #ifndef __CTHUNK_H__
 #define __CTHUNK_H__
 
-#include "platform/source/CThunkBase.h"
+#include "platform/internal/CThunkBase.h"
 
 /**
  * Class for created a pointer with data bound to it
