+++ platform/internal/CThunkBase.h
